facturen index fname and index fix
fixed the following issues:
in model persoon, the fillables were not capitalized properly.
in model persoon, the HasOne was not defined properly, and capitalized.
when creating a factuur through factory, non existing column klant_id was attempted to be filled
at that point it was changing up the properties in index.blade
Most of these issues stem from inproper naming conventions. Please take due care of using it properly